Spanish Tranlation/Location
I have a question though, what does the " ' ' " stand for in "Wrench"? Does it affect the word inside it? I couldn't find the string that translates the wrench into spanish. Is there such thing as a modifier for the item name of "Wrench"? Checking onto the item prototypes, there's not one.
Did you miss it?

Anyways, here's the translation, please let me know if you have any doubts.
Greets!